Steven multiple windows (#381)
Fixes #321

## Changes Made

- In the CanvasPlusGroundApp: each window instance should have its own
`NavigationModel` instance in order to be independent of each other's
navigation state, thus removing the global shared `NavigationModel`
instance
- Created a new window group for a newly defined data type
`FocusWindowInfo` needed to retain necessary info about which sub-page
of which course the user is trying to open a new focus window for
- Created a dedicated custom `CourseContextMenu` view modifier attached
to each `NavigationLink` of each course page in CourseView in order to
push an instance of data type `FocusWindowInfo`. This view modifier will
call `@Environment(\.openWindow)` along with information about the
current course and the subpage the user is in in order to create a new
instance of the Focus Window Group.

## Screenshots


https://github.com/user-attachments/assets/70037dd8-ecae-4aa1-968c-085e5706175d



## Checklist
- [x] I have implemented all requirements for this PR and its
accompanying issue.
- [x] I have verified my implementation across all edge cases.
- [x] I have ensured my changes compile on macOS & iOS.
- [x] I have resolved all SwiftLint warnings.