Implement Course Home View (#391)
Fixes #368

## Changes Made

- Check the `CourseDefaultView` type to conditionally display different
home page views
  - For `modules` and `assignments`, reuse the existing views
- For `wiki`, create and use a new `GetFrontPageRequest` to get the
front page as a `Page`. Use `PageView` to display this
- For `syllabus`, use the `syllabusBody` property of `Course` and then
`HTMLView` to display the page.

- Fix a navigation destination runtime warning that was occurring due to
changes in #381. (cc @ecoist-ste)

Note: For `activity`, we will need to support the Course Activity Stream
which will require more work. I don't think many courses use this right
now, so we can aim to get this done as part of the featurework milestone
instead.

## Screenshots (if applicable)

![Screen Recording 2025-09-06 at 11 24
44 AM](https://github.com/user-attachments/assets/b377be98-7254-4a6f-a51f-dfc3e26dbb93)

## Checklist
- [x] I have implemented all requirements for this PR and its
accompanying issue.
- [x] I have verified my implementation across all edge cases.
- [x] I have ensured my changes compile on macOS & iOS.
- [x] I have resolved all SwiftLint warnings.