Multi-window support for individual items (#395)
Fixes #389 

## Changes Made

- Add the ability to open specific course materials, including the
courses themselves, in their own windows.
- Introduce `CodableDestination`: since `NavigationModel.Destination` is
not codable, we need to create a version of it that is for the
`openWindow` function.
- Refactor `CourseContextMenu` to `NewWindowButton` which is a reusable
button that can open a given item in a new window.
- `FocusWindowView` is refactored to load the specific item upon appear.
This is needed since only the id of the item is passed in, as opposed to
the item itself. It also is required when these windows are restored
when the app is relaunched.
- Add a `GetFolderRequest`, which allows retrieving and caching specific
folders. Needed if a folder is opened in a new window.

## Screenshots (if applicable)

![Screen Recording 2025-09-10 at 10 51
05 AM](https://github.com/user-attachments/assets/61bd6e08-1478-4a5e-906c-06bcaa1a6ca0)

## Checklist
- [x] I have implemented all requirements for this PR and its
accompanying issue.
- [x] I have verified my implementation across all edge cases.
- [x] I have ensured my changes compile on macOS & iOS.
- [x] I have resolved all SwiftLint warnings.