[IGC v1] Introduce Intelligence Foundations + Support (#419)
~~Fixes~~ Work towards #249

This PR introduces a set of foundational tools which will be used by the
Intelligent Grade Calculator (IGC), including the intelligence and file
picker structure. The changes include a debug playground which uses this
intelligence to output a set of assignment categories and weights.
Please use this to test the functionality.

## Changes Made

- Introduce `GradeCalculatorIntelligenceService` which returns a set of
assignment groups which can later be used by the Grade Calculator
directly. The prompt and preprocessing architecture will need to
continuously be iterated on, and we should mark this feature as "Beta"
when released initially.
- `IGCPlayground` is a debug view accessible on macOS, which can be used
to test the intelligence.
- There will be 2 final ways for user's to select a syllabus file:
`CourseItemPicker` and a regular file upload. Both are accessible
through the debug view.

## Screenshots (if applicable)


https://github.com/user-attachments/assets/8a7dd48e-f774-4f98-84fc-5dcefd4c2b40

## Checklist
- [x] I have implemented all requirements for this PR and its
accompanying issue.
- [x] I have verified my implementation across all edge cases.
- [x] I have ensured my changes compile on macOS & iOS.
- [x] I have resolved all SwiftLint warnings.