[Widgets] Implement on-demand widget refresh (#457)
Fixes --

## Changes Made

- As a part of #451, there was an issue where widgets would not refresh
when onboarding was complete. As a part of this PR, data sources now do
not require a `refreshTrigger`. Instead, this is part of the global
`shared` instance of `WidgetContext`, allowing any data source to
refresh widgets on-demand using `requestToRefreshAllWidgets` or
`requestToRefreshWidget(widget:)`.
- This is called when `CourseManager` fetches the latest courses, fixing
the earlier bug.
- We should also implement this in other data sources (like
announcements and todos), will do as part of another PR.

## Screenshots (if applicable)

## Checklist
- [x] I have implemented all requirements for this PR and its
accompanying issue.
- [x] I have verified my implementation across all edge cases.
- [x] I have ensured my changes compile on macOS & iOS.
- [x] I have resolved all SwiftLint warnings.